refactor(data-layer): consolidate schema, drop migration runners
No deployed instances to migrate from, so the staged-migration machinery is
deleted. The full final-state schema lives in two SQL files (one per backend)
and is applied once on startup via idempotent `CREATE ... IF NOT EXISTS`.

- New: db/sqlite-schema.sql + db/schema-applier.ts (applySchema)
- New: repository/postgres/schema.sql + repository/postgres/schema-applier.ts
       (applyPostgresSchema)
- Deleted: db/migrate.ts, migrations/index.ts, all migrations/*.sql (25 files);
           repository/postgres/migrate.ts, postgres/migrations/*.sql (4 files)
- Deleted: _migrations tracking table and the version + name-based runner
- Updated callers: persistence.ts, test-support/test-db.ts, postgres tests
- Rewrote db/schema-applier.test.ts to cover applySchema invariants
