+++ packages/api-gateway/src/app/persistence.ts
@@ -8,10 +8,9 @@
  *   - SQLite (default) — single-file DB under DATA_DIR / SQLITE_PATH.
  *   - Postgres-hybrid — `DATABASE_URL=postgres(ql)://...` swaps the W2
  *     instance-config repos (LLM, datasources, notification channels) to
- *     Postgres while everything else stays on SQLite. Postgres migrations
- *     run via `applyPostgresInstanceMigrations`; we await them now (was
- *     fire-and-forget) so the gateway never serves a request against an
- *     unmigrated schema.
+ *     Postgres while everything else stays on SQLite. The Postgres schema
+ *     is applied awaited (not fire-and-forget) so the gateway never serves
+ *     a request against an unbuilt schema.
  *
  * Any non-Postgres `DATABASE_URL` value is logged at warn and ignored —
  * historically that env var gated an in-memory mode that W2 deleted.
@@ -20,13 +19,12 @@
 import {
   createSqliteClient,
   createSqliteRepositories,
-  ensureSchema,
-  applyNamedMigrations,
+  applySchema,
   createDbClient,
   PostgresInstanceConfigRepository,
   PostgresDatasourceRepository,
   PostgresNotificationChannelRepository,
-  applyPostgresInstanceMigrations,
+  applyPostgresSchema,
 } from '@agentic-obs/data-layer';
 import type { SqliteRepositories } from '@agentic-obs/data-layer';
 import { createLogger } from '@agentic-obs/common/logging';
@@ -52,8 +50,7 @@ function isPostgresUrl(
 
 function buildSqlite(): { repos: SqliteRepositories; sqliteDb: SqliteClient } {
   const sqliteDb = createSqliteClient({ path: dbPath() });
-  ensureSchema(sqliteDb);
-  applyNamedMigrations(sqliteDb);
+  applySchema(sqliteDb);
   return { repos: createSqliteRepositories(sqliteDb), sqliteDb };
 }
 
@@ -63,8 +60,8 @@ async function buildPostgresHybrid(
   const base = buildSqlite();
   const pg = createDbClient({ url });
   // Awaited (was fire-and-forget) — boot must not race instance-config
-  // queries against an unmigrated schema.
-  await applyPostgresInstanceMigrations(pg);
+  // queries against an unbuilt schema.
+  await applyPostgresSchema(pg);
   return {
     sqliteDb: base.sqliteDb,
     repos: {
