+++ packages/data-layer/src/db/index.ts
@@ -4,5 +4,4 @@ export type { DbClient, DbClientOptions } from './client.js';
 export * as sqliteSchema from './sqlite-schema.js';
 export { createSqliteClient } from './sqlite-client.js';
 export type { SqliteClient, SqliteClientOptions } from './sqlite-client.js';
-export { ensureSchema, migrate, applyNamedMigrations, splitSqlStatements } from './migrate.js';
-export { loadSqlMigrations, type SqlMigration } from '../migrations/index.js';
+export { applySchema, splitSqlStatements } from './schema-applier.js';
