+++ packages/data-layer/src/repository/postgres/instance-config.test.ts
@@ -14,7 +14,7 @@
 import { describe, it, expect, beforeAll, afterAll, beforeEach } from 'vitest';
 import { sql } from 'drizzle-orm';
 import { createDbClient, type DbClient } from '../../db/client.js';
-import { applyPostgresInstanceMigrations } from './migrate.js';
+import { applyPostgresSchema } from './schema-applier.js';
 import { PostgresInstanceConfigRepository } from './instance-config.js';
 
 const PG_URL = process.env['POSTGRES_TEST_URL'];
@@ -31,7 +31,7 @@ describeIfPg('PostgresInstanceConfigRepository', () => {
     process.env['SECRET_KEY'] =
       prevSecret ?? 'test-secret-key-for-instance-config-repositories-xxxxxxxx';
     db = createDbClient({ url: PG_URL! });
-    await applyPostgresInstanceMigrations(db);
+    await applyPostgresSchema(db);
   });
 
   afterAll(() => {
