+++ packages/data-layer/src/test-support/test-db.ts
@@ -1,9 +1,9 @@
 import { createSqliteClient, type SqliteClient } from '../db/sqlite-client.js';
-import { migrate } from '../db/migrate.js';
+import { applySchema } from '../db/schema-applier.js';
 
 /**
- * Open a fresh in-memory SQLite database, run *all* migrations (legacy V1..V4
- * plus every named auth/perm migration), and return the wired SqliteClient.
+ * Open a fresh in-memory SQLite database, apply the consolidated schema, and
+ * return the wired SqliteClient.
  *
  * Intended for unit/integration tests. Every new repository test in this
  * wave calls this to get a clean DB.
@@ -14,6 +14,6 @@ import { migrate } from '../db/migrate.js';
  */
 export function createTestDb(): SqliteClient {
   const db = createSqliteClient({ path: ':memory:', wal: false });
-  migrate(db);
+  applySchema(db);
   return db;
 }
